+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed 🔧
 
 * Renamed `FontFamily::VariableWidth` to `FontFamily::Proportional`.
+* Remove `pixels_per_point` from `FontDefinitions`.
 
 ### Fixed 🐛